% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installdic.R
\name{install_dic}
\alias{install_dic}
\title{Install Mecab-Ko-Dic in Linux and Mac OSX.}
\usage{
install_dic()
}
\value{
None. The function will halt when the current operation system is not Linux or Mac OSX, or Mecab-Ko-Dic is installed already.

See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.
}
\description{
\code{install_dic} installs Mecab-Ko-Dic.
}
\details{
This code checks and installs Mecab-Ko-Dic in Linux and Mac OSX. This is essential for using custom-defined user dictionary. Installing Mecab-Ko-Dic needs system previleges, because it uses `make install` to build from source and install it to system.
}
\examples{
\dontrun{
install_dic()
}

}
