% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mecab}
\alias{install_mecab}
\title{Install mecab-ko-msvc and mecab-ko-dic-msvc}
\usage{
install_mecab(mecabLocation)
}
\arguments{
\item{mecabLocation}{a directory to install Mecab-Ko-MSVC and Mecab-Ko-Dic-MSVC.}
}
\value{
None. The function will halt when the current operation system is not Windows, or /mecabLocation/mecab.exe exists.

See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.
}
\description{
\code{install_mecab} installs Mecab-Ko-MSVC and Mecab-Ko-Dic-MSVC.
}
\details{
This code checks and installs Mecab-Ko-MSVC and Mecab-Ko-Dic-MSVC in user specified directory. Windows only.
}
\examples{
\dontrun{
install_mecab("D:/Rlibs/mecab")
}

}
