\name{nouns}
\alias{nouns}
\title{Noun extractor by mecab-ko}
\description{
  The \code{nouns} function is used for noun extraction of Korean sentence.
}
\details{
  
  Noun extraction by mecab-ko and mecab-ko-dic.
  See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.

}
\references{
  \itemize{
  	\item \href{http://eunjeon.blogspot.com}{Eunjeon project}
  	\item \href{https://github.com/Pusnow}{Wonsup Yoon}, mecab-ko VC++ builds
  }
}
\author{Junhewk Kim}
\usage{
  nouns(phrase)
}
\arguments{
  \item{phrase}{input character vector}
}
\keyword{tagger}
\keyword{nlp}
\keyword{korean}