% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RmecabKo-package}
\alias{RmecabKo}
\alias{RmecabKo-package}
\title{Korean text analysis with MeCab}
\description{
`RmecabKo` provides Korean-specific extraction, tokenization, and n-grams.
The native MeCab interface is supplied by [RcppMeCab::RcppMeCab].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/junhewk/RmecabKo}
  \item Report bugs at \url{https://github.com/junhewk/RmecabKo/issues}
}

}
\author{
\strong{Maintainer}: Junhewk Kim \email{junhewk.kim@gmail.com}

Authors:
\itemize{
  \item Junhewk Kim \email{junhewk.kim@gmail.com}
}

}
\keyword{internal}
