% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nouns.R
\name{words}
\alias{words}
\title{Words extractor by mecab-ko}
\usage{
words(phrase)
}
\arguments{
\item{phrase}{Character vector.}
}
\value{
List of full morphemes will be returned.

See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.
}
\description{
\code{words} returns full morphemes extracted from Korean phrases.
}
\details{
It is based on Mecab-Ko POS classification.
}
\examples{
\dontrun{
words(c("Some Korean Phrases"))
}
}
