% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nouns.R
\name{nouns}
\alias{nouns}
\title{Noun extractor by mecab-ko}
\usage{
nouns(sentence, sys_dic = "", user_dic = "", parallel = FALSE)
}
\arguments{
\item{phrase}{A character vector or character vectors.}
}
\value{
List of nouns will be returned. Element name of the list are original phrases.

See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.
}
\description{
\code{nouns} returns nouns extracted from Korean phrases.
}
\details{
Noun extraction is used for many Korean text analysis algorithms. The function coerces input to UTF-8.
}
\examples{
\dontrun{
nouns(c("Some Korean Phrases"))
}

}
