% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{pos}
\alias{pos}
\title{POS tagging by mecab-ko}
\usage{
pos(phrase)
}
\arguments{
\item{phrase}{Character vector.}
}
\value{
List of POS tagged morpheme will be returned. Element name of the list are original phrases.

See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.
}
\description{
\code{pos} returns part-of-speech (POS) tagged morpheme of Korean phrases.
}
\details{
This is a basic function of part-of-speech tagging by mecab-ko.
}
\examples{
\dontrun{
pos(c("Some Korean Phrases"))
}

}
