\name{pos}
\alias{pos}
\title{POS tagging by mecab-ko}
\description{
  The \code{pos} function is used for POS tagging of Korean sentence.
}
\details{
  
  POS tagging by mecab-ko and mecab-ko-dic.
  See examples in \href{https://github.com/junhewk/RmecabKo}{Github}.

}
\references{
  \itemize{
  	\item{\href{http://eunjeon.blogspot.com}{Eunjeon project}}
  	\item{\href{https://github.com/Pusnow}{Wonsup Yoon}, mecab-ko VC++ builds}
  }
}
\author{Junhewk Kim}
\usage{
  pos(phrase)
}
\arguments{
  \item{phrase}{input character vector}
}
\keyword{tagger}
\keyword{nlp}
\keyword{korean}