% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords_ko}
\alias{stopwords_ko}
\title{Stopwords for Korean Morpheme}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 433 rows and 1 columns.}
\usage{
data(stopwords_ko)
}
\description{
Stopwords dataset developed based on results of scraped articles.
It is structured as data frame and tibble (dplyr), so if you want use this stopwords set in the tokenizer, call it directly.
}
\examples{
\dontrun{
# Basic usage: using it combined with tidyverse and tidytext package
library(tidyverse)
library(tidytext)
df \%>\% unnest_tokens(word, txt, token = token_words) \%>\% 
    anti_join(stopwords_ko)

# Use stopwords in token_ngram
token_ngram(txt, n = 2L, stopwords = stopwords_ko$word)
}

}
\keyword{datasets}
